docs: add doc for api utils
ref: MANAGER-11306
